seperating Header and Footer components from the Layout component
