+++ src/components/Layout.jsx
@@ -1,0 +1,43 @@
+import React from "react"
+import PropTypes from "prop-types"
+import { useStaticQuery, graphql, Link } from "gatsby"
+
+import Header from "./Header.jsx"
+import Footer from "./Footer.jsx"
+import "./layout.css"
+
+const Layout = ({ children }) => {
+  const data = useStaticQuery(graphql`
+    query SiteTitleQuery {
+      site {
+        siteMetadata {
+          title
+        }
+      }
+    }
+  `)
+
+  return (
+    <>
+      <div style={{ margin: `3rem auto`, maxWidth: 650, padding: `0 1rem` }}>
+        <Header siteTitle={data.site.siteMetadata.title} />
+      </div>
+      <div
+        style={{
+          margin: `0 auto`,
+          maxWidth: 960,
+          padding: `0 1.0875rem 1.45rem`,
+        }}
+      >
+        <main>{children}</main>
+        <Footer />
+      </div>
+    </>
+  )
+}
+
+Layout.propTypes = {
+  children: PropTypes.node.isRequired,
+}
+
+export default Layout
